Update Group4SDP/src/computer/ControlGUI2.java
Changed to GUI layout, fixed the nonworking fields, buttons etc.
Removed the timers from actions - they interfere with other actions
(still happens in a lot of cases). It is easier to manually stop the robot,
which happens in most cases anyway.
Also implemented graceful quit for the GUI, even if the robot is not connected.